Subjects Added
feat: Add Environmental Science subject and update subject pages

This commit adds the Environmental Science subject to the list of subjects in the constants file. It also updates the subject pages for MA24101 and ES24101 to include new sections for PDFs and PYQs. Additionally, it adds a new section for Best YouTube Videos in the subject pages.